+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 HumHub Changelog
 ================
 
-1.17.0-beta.5 (Unreleased)
----------------------------------
+1.17.1 (Unreleased)
+-------------------
+- Fix #7377: Configured Redis cache key prefix was overwritten by the default value
+- Enh #7383: Improve SelfTest for Base URL 
+
+1.17.0 (January 13, 2025)
+-------------------------
+
+See also: HumHub 1.17 - [Release Notes](https://docs.humhub.org/docs/about/releasenotes/release_notes_1_17) and [Changelog](https://github.com/humhub/humhub/blob/master/CHANGELOG.md)
+
+Info: The minimum PHP version is now `PHP 8.1`!
+
 - Fix #7365: `DeviceDetectorHelper::isMobile()` and `DeviceDetectorHelper::isTablet()` when no user agent
+- Fix #7376: `humhub\helpers\ArrayHelper::flatten()` not compatible Yii base ArrayHelper
 
 1.17.0-beta.4 (December 24, 2024)
 ---------------------------------